+++ app/Http/Controllers/CursoController.php
@@ -17,7 +17,7 @@ class CursoController extends Controller
     public function index()
     {
         //
-        Gate::authorize('curso');
+        Gate::authorize('read-curso');
         return view('pages.cursos.index');
     }
 
@@ -40,6 +40,7 @@ class CursoController extends Controller
     public function store(StoreCursoRequest $request)
     {
         //
+        Gate::authorize('insert-curso');
         $dados =Curso::create($request->all());
        
         return response()->json($dados);
@@ -54,6 +55,7 @@ class CursoController extends Controller
     public function show(Curso $curso)
     {
         //
+        Gate::authorize('read-curso');
         return response()->json($curso->all());
     }
 
@@ -78,6 +80,7 @@ class CursoController extends Controller
     public function update(Request $request, Curso $curso)
     {
         //
+        Gate::authorize('update-curso');
         if($request->ajax()){
             $curso->find($request->input('pk'))->update([$request->input('name') => $request->input('value')]);
             return response()->json(['success' => true]);
@@ -93,6 +96,7 @@ class CursoController extends Controller
     public function destroy($id)
     {
         //
+        Gate::authorize('delete-curso');
         Curso::find($id)->delete();
         return response()->json(['success' => true]);
     }
